+++ android/src/main/java/com/jhotadhari/reactnative/mapsforge/vtm/react/views/MapFragment.java
@@ -179,7 +179,6 @@ public class MapFragment extends Fragment {
 		propHgtReadFileRate = hgtReadFileRate;
 		propEmitsMapEvents = emitsMapEvents;
 		propEmitsHardwareKeyUp = emitsHardwareKeyUp;
-
     }
 
 	public void updateHardwareKeyListener( List<String> emitsHardwareKeyUp ) {